+++ app/api/responses.js
@@ -1,0 +1,12 @@
+const { get } = require('./base')
+const { responsesApi: config } = require('../config/api')
+
+const baseUrl = config.baseUrl
+
+const getResponses = async (docId) => {
+  return get(`${baseUrl}/responses/${docId}`)
+}
+
+module.exports = {
+  getResponses
+}
