--- app/config/documents-api.js
@@ -1,17 +1,0 @@
-const Joi = require('joi')
-
-const schema = Joi.object({
-  baseUrl: Joi.string().required()
-})
-
-const config = {
-  baseUrl: process.env.DOCUMENTS_API_BASE_URL
-}
-
-const { error, value } = schema.validate(config, { abortEarly: false })
-
-if (error) {
-  throw new Error(`The message queue config is invalid. ${error.message}`)
-}
-
-module.exports = value
